Remove docker-compose.yml
We have built a docker based local environment specifically for working locally. The docker-compose.yml we're removing is an attempt at getting the Docker image the current Dockerfile builds running as something we can work with locally. Our new environment does it better so we don't want this file hanging around causing confusion.