Changes to get back to AWS (#724)
In early 2022 the SROC team were coming to the end of development of the [SROC Charging Module API (CHA)](https://gitub.com/DEFRA/sroc-chargig-module-api). It was developed for use by another team that was not yet ready to deploy their side of the project to production.

So, the SROC dev team, looking for ways to continue improving things for the business opted to move the TCM from being EC2-based (deployed to a server) to being ECS-based (deployed as a Docker image).

The team had developed the CHA to be Docker-based, and all new services in Defra were expected to be built on Docker. This would allow the service to take advantage of a cheaper, future-proof deployment model and environment. Being an existing service that was internal only it was also selected as a pilot for web-ops to help develop a 'standardised' ECS setup and infrastructure.

In [Add Docker support to the project](#515) the breaking change was made. From that point onwards the project would expect to be built as an image and run as a container. It could no longer be deployed to EC2 using the existing methods.

Fast forward to May 2023 and that ideal was still to be realised. The SROC delivery team were no more, having moved on to other projects. When they could they were still pitching in with the pilot. But that had grown in scope and involved many more parties so nothing tangible was ready for use. Then crisis happened. The version of the TCM deployed to `production` was using [SendGrid](https://sendgrid.com/) to handle email sending. But failures in users receiving them, or losing them in junk (not SendGrid's fault by the way, we spotted too late that the original developers had never properly configured SendGrid and the AWS DNS) meant it needed sorting.

It just so happens that one of the other improvements the team made before leaving the project was to switch it from SendGrid to [GOV.UK Notify](https://www.notifications.service.gov.uk/). But the switch was implemented _after_ the breaking change to Docker. So, we branched off the [version (3.2.1)](https://github.com/DEFRA/sroc-tcm-admin/releases/tag/v3.2.1) deployed to production, replicated the change, and deployed that 'hotfix' branch.

That was where the TCM was left with those of us keeping the lights on still hopeful that migration to ECS would happen. We now have to accept that will never happen for 2 reasons

- The TCM needs an **Annual billing upload** feature that was [stripped as unneeded](#416) re-implemented and working in time to be tested for April 2024 to support changes in EPR
- Plans are already in place to move 2 of the 3 regimes TCM supports off the system in Q1 and Q2 of this year. Then work will begin on the last one

The team made several improvements to the TCM in the last days of the project. But they are all locked behind the breaking change.

These changes are focused on getting the TCM deployed and running using the existing Capistrano scripts and AWS EC2 infrastructure. We've also created a local dev environment which will be used for working on the TCM going forward which will necessitate some changes.

---

- Remove docker-compose.yml

We have built a docker-based local environment specifically for working locally. The docker-compose.yml we're removing is an attempt at getting the Docker image of the current Dockerfile builds running as something we can work with locally. Our new environment does it better so we don't want this file hanging around causing confusion.

- Update README to point to new dev enviroment

We've built a new environment that makes working locally with the project easier. So, new users shouldn't use the environment the `docker-compose.yml` was creating.

This updates the README to point folks that way and removes the references to local VSCode tasks.

- Remove .vscode

This will now be managed by the **tcm-dev-environment** we've built.

- Add passenger gem back in

This is what is used to run the app on our AWS EC2 environments. We're adding it back in so we can get the app running again.

- Move airbrake into standard gems

The **tcm-dev-environment** has an instance of Errbit so it should be fine to have Airbrake loaded even when we're not in production mode.

- Resolve asset precompile error in AWS

When running the deployment it is erroring during the asset precompile step.

```
$HOME/.rbenv/bin/rbenv exec bundle exec rake assets:precompile

rake aborted!
01 Uglifier::Error: Unexpected token: keyword (const). To use ES6 syntax, harmony mode must be enabled with Uglifier.new(:harmony => true).
```

The first [stackoverflow post](https://stackoverflow.com/a/57943222/6117745) we found suggested updating `config/environments/production.rb` to use `config.assets.js_compressor = Uglifier.new(harmony: true)`.

- Remove development stage from Dockerfile

We'll not be using it any more and it is our convention not to include code we don't maintain.

- Fix getting the commit sha

We updated getting the commit sha to only work in Docker. We now want it to work in our local dev environment, AWS, and the docker image we're still triggering after each change to `main`.