Add back in Annual billing upload functionality
When last working on the project as part of work to remove redis and [resque](https://github.com/resque/resque) dependencies we [removed Annual billing upload functionality](https://github.com/DEFRA/sroc-tcm-admin/pull/416).

At the time we were told this was a defunct feature and only needed when the service went live. Removing it saved us updating the code to no longer use resque.

Problem is, we need it back!

As part of getting ready for EPR in 2024 the regimes that use the TCM need to load in 20,000 new activities and their charge categories. This was what the 'Annual billing upload' was built for so they'd really like it back.

This change adds it back into the UI and the service. We can't just copy the code we deleted though as we now need to do the work to strip the dependence on resque. Fortunately, there were other jobs like the file import and export processes that have already gon e through this.