Add back in Annual billing upload functionality (#726)
When last working on the project as part of work to remove redis and [resque](https://github.com/resque/resque) dependencies we [removed Annual billing upload functionality](https://github.com/DEFRA/sroc-tcm-admin/pull/416).

We were told this was a defunct feature and only needed when the service went live. Removing it saved us from updating the code to no longer use Resque.

The problem is, we need it back!

As part of getting ready for EPR in 2024, the TCM regimes need to load 20,000 new activities and their charge categories. This was what the 'Annual billing upload' was built for so they'd like it back.

This change adds it back into the UI and the service. We can't just copy the code we deleted though as we now need to do the work to strip the dependence on resque. Fortunately, other jobs like the file import and export processes have already gone through this.